[DataMapper] Code Cleanup, move some API out of DataMapper.hpp to keep that file cleaner
